backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
redis==6.4.0
celery==5.5.3
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
scikit-learn==1.7.1
xgboost==3.0.4
numpy==2.3.2
pandas==2.3.1
pytest==8.4.1
cryptography==45.0.6
